Fixed flickering, finally!
So it seems that the CWBackPixel flag was causing the trouble when the window is being resized, not even 50% sure, but the reason could have been the way X server handles the configure notify/expose event with the CWBackPixel enabled and it results in repainting or maybe deleting the window's foreground/plane (when there is content)? The documentation for `XCreateWindow` is not really clear on what results on enabled/disabled flags on the valuemask and I didn't find any additional info on the matter.