Following next part in the series (video 6) and adding render context class for drawing
It's suppose to be fast drawing, but it's not really that fast?

I wrote it in a different matter, that isn't allocating a flat array + extending the bitmap class, but instead creating a standalone class with allocating an array of namedtuples and mapping pixels to the display's surface object. This difference could be why it's slower than expected, but atm there's no benchmark numbers from the java version. Update: looks like the java version takes about the same amount of time for drawing (1.4ms).

In debug avg draw time is about 6ms, and release avg is 1.3ms.