Added Vector4 obj, replacing Vertex, not fully tested yet and needs some docs, thou it does compile and run with the Vector4 changes
First attempt was trying to copy StaticArray for similar semantics, but it wasn't working out very well with regards to allocation of the StaticArray when it came to the number of axes when init'ing the object with the type and number (num of axis). It will be easier and smaller to just keep defined axes in the object as variables instead. Better for now to continue progressing forward instead fiddling about with such a minor thing.

Note: should still refer to the name Vertex for vertices rather than using the Vector4 obj? could have just update the alias instead?

In the future this object could use crystal macros to also make 2 and 3 axis versions if needed.