Added Shale::Edge class and refactored Shale::RenderCtx
The Shale::Edge class is used to find the edges of 2 vertices and works work with the Shale::RenderCtx scan methods to find the start and end x positions for each y coordinate.

Note, Shale::Edge was defined as a class instead of a struct because it needed to keep its state when iterating, other wise it doesn't map out the pixels correctly. I think this is due to crystal with a class being allocated on the heap and passed by reference to other functions compared to structs that are allocated on the stack and passed by value, so when it was a struct, the Shale::Edge#x value wasn't correct on the second call to Shale::RenderCtx#scan_edges.
Not sure if there's a way to pass structs by reference in crystal, but for now should be fine as a class, though i don't know what performance differences there are in this scenario.